integration du cycle de question/reponse dynamique

+++ devoxx-game/src/main/java/fr/soat/devoxx/game/model/Question.java
@@ -15,7 +15,7 @@ public class Question {
 
 	@Id
 	@Column(name = "id")
-	Integer idQuestion;
+	Long idQuestion;
 	
 	@Column(name = "LABEL")
 	String questionLabel;
@@ -26,11 +26,11 @@ public class Question {
 	@OneToOne
 	QuestionChoice goodChoice;
 
-    public Integer getIdQuestion() {
+    public Long getIdQuestion() {
         return idQuestion;
     }
 
-    public void setIdQuestion(Integer idQuestion) {
+    public void setIdQuestion(Long idQuestion) {
         this.idQuestion = idQuestion;
     }
